Add files via upload
Added quizzes three and four, which tested knowledge of if-else statements and the input function. I also adjusted some functions which were causing issues with returning to frames after exiting to Home. 'Tutorials' have been replaced with guides, of which there are currently the beginnings of two. Guide one and five are incomplete with only the backgrounds currently in place. Small style changes were made such as the font swap of the home buttons.