Add files via upload
Completed the content of the app with all ten quizzes and guides being implemented. The dynamic back button was swapped out due to unnecessary complexity for the user, there are now menu buttons and the back button takes you to the previous quiz / guide. The help buttons for all ten quizzes were also made with additional functions being added to show and hide their frames, along with a toggle system that allows the same button to close/open the help frame.